+++ ArcdpsLogManager/CHANGELOG.md
@@ -16,6 +16,10 @@ This is the full changelog of the arcdps Log Manager.
 - Fixed log detail pane not appearing for selected logs processed with Log Manager versions 1.3 and older.
 - Fixed a stray debug data section appearing when multiple logs were selected even if debug data wasn't enabled
 
+#### EVTC Inspector notes
+- Fixed buff extension events not being distinguished from normal buff applies
+- Added buff instance ids to buff applies / extends
+
 ## Log Manager v1.9
 #### New features
 - Added support for Silent Surf CM and NM